Update user_bucket owner in SQL script for data purge process
This commit updates the SQL script responsible for finding purge-eligible patrons and writing INSERT statements to change the owner of the user_bucket. The owner is now set to a specific ID, 1916108, instead of the generic library ID.

Release-Note: Update data purge script to set a fixed owner for user_buckets.

Signed-off-by: Ian Skelskey <ianskelskey@gmail.com>